backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20

# Database
sqlmodel>=0.0.22
sqlalchemy>=2.0.45
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Pydantic
pydantic>=2.10.0
pydantic-settings>=2.6.0

# HTTP and utilities
httpx>=0.27.0
python-dotenv>=1.0.0

# OpenAI Agents SDK (required for Phase III)
openai-agents>=0.6.5
openai>=2.15.0
